Phase 4-6: mypy clean, pre-commit hooks, CI pipeline, 46 frontend tests
- Fix all 25 mypy errors: TYPE_CHECKING imports for model forward refs,
  json.loads type annotations, async_sessionmaker, event_bus UUID cast
- Add SQLAlchemy mypy plugin to pyproject.toml
- Add per-file-ignores for test S105/S106/S108 false positives
- Exclude migrations from ruff
- Add mypy hook to .pre-commit-config.yaml (all 4 hooks pass)
- Make CI mypy blocking, add --cov-fail-under=95 coverage gate
- Add AddressInput (8 tests) and dashboard store (11 tests) frontend tests
- Total: 657 backend tests (100% coverage), 46 frontend tests

+++ backend/nile/services/chain_service.py
@@ -20,7 +20,7 @@ def _load_idl(program_name: str) -> dict | None:
     if not idl_path.exists():
         logger.warning("IDL not found: %s", idl_path)
         return None
-    return json.loads(idl_path.read_text())
+    return json.loads(idl_path.read_text())  # type: ignore[no-any-return]
 
 
 def validate_solana_address(address: str) -> bool:
@@ -229,7 +229,7 @@ class SolanaChainService:
             exponent = struct.unpack_from("<i", data, 20)[0]
             price = struct.unpack_from("<q", data, 208)[0]
 
-            return price * (10**exponent)
+            return float(price * (10**exponent))
         except Exception:
             logger.exception("Failed to get SOL price from Pyth")
             return None
